Add support for viewporter

+++ wlcs_anvil/src/renderer.rs
@@ -131,7 +131,7 @@ impl Frame for DummyFrame {
     fn render_texture_from_to(
         &mut self,
         _texture: &Self::TextureId,
-        _src: Rectangle<i32, Buffer>,
+        _src: Rectangle<f64, Buffer>,
         _dst: Rectangle<f64, Physical>,
         _damage: &[Rectangle<f64, Physical>],
         _src_transform: Transform,
